Restore out.format to its former value
out.format needs to be set for knit_theme$get(thm) to work.

However, when building a pkgdown article based on a vignette
that uses BiocStyle::pdf_document, the out.format is forced
to be html, and hardcoding this knit option clashes with
pkgdown.

By setting and restoring the output format, things keep
working as expected